+++ src/components/DownloadButton.vue
@@ -13,7 +13,6 @@ import {
 })
 
 export default class DownloadButton extends Vue {
-  @Prop({ type: Function, required: true }) readonly callRefrehContext!: () => void;
   @Prop() readonly context!: ProjectManagerContext;
   @Prop() readonly projectManagerBackendService!: ProjetManagerBackendService;
   @Prop() readonly module!: Module;
@@ -23,25 +22,25 @@ export default class DownloadButton extends Vue {
   @Prop() readonly text: string | undefined = undefined;
   isActive = false;
 
-  @Watch('projectManagerBackendService', { immediate: true, deep: true })
+  @Watch('projectManagerBackendService', {immediate: true, deep: true})
   onContextChange(newValue: ProjetManagerBackendService, oldValue: ProjetManagerBackendService) {
     this.updateIsActive()
   }
+
   async created() {
     this.updateIsActive()
   }
 
-  updateIsActive(){
+  updateIsActive() {
     this.projectManagerBackendService.isModuleActionActive(this.module, this.action).then(result => this.isActive = result)
   }
 
   downloadFile(): void {
-    const params = new Map<string,unknown>();
-    if (this.filename){
+    const params = new Map<string, unknown>();
+    if (this.filename) {
       params.set('filename', this.filename);
     }
     this.projectManagerBackendService.downloadFile(this.module, this.action, this.context, params).then(httpResponse => {
-      this.callRefrehContext();
       this.updateIsActive();
     });
   }
@@ -51,12 +50,12 @@ export default class DownloadButton extends Vue {
 
 <template>
   <div v-if="isActive">
-    <span v-if="text">{{text}}</span>
+    <span v-if="text">{{ text }}</span>
     <div v-if="!iconClass">
       <button @click="downloadFile">Download</button>
     </div>
     <div v-if="iconClass">
-      <i class="bi bi-download" @click="downloadFile" ></i>
+      <i class="bi bi-download" @click="downloadFile"></i>
     </div>
   </div>
 </template>